fix: suspense dialog content
