+++ apps/web/src/components/ui/provider/FitnessDataModalProvider/FitnessDataModalProvider.tsx
@@ -1,14 +1,9 @@
 import type { ModalContextType } from './FitnessDataModalContext';
 import type { ReactNode } from 'react';
-import { Button, Modal, ModalBody, ModalContent, ModalFooter, ModalHeader, ScrollShadow } from '@heroui/react';
-import { useEffect, useMemo, useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import { useLazyGetExerciseFinishHistory } from '@hooks/apollo/Exercise';
-import { useFitnessFragment } from '@hooks/apollo/Fitness';
-import StateRender from '@utils/StateRender';
-import DisplayFitnessFinishHistory from './DisplayExerciseFinishHistory';
+import { Modal, ModalContent } from '@heroui/react';
+import { Suspense, useState } from 'react';
+import FitnessDataModalContent from './FitnessDataModal';
 import { ModalContext } from './FitnessDataModalContext';
-import FitnessPreviewVideo from './FitnessPreviewVideo';
 
 interface ExerciseDataInfoModalProps {
   children: ReactNode
@@ -17,116 +12,28 @@ interface ExerciseDataInfoModalProps {
 export default function FitnessDataModalProvider({
   children
 }: ExerciseDataInfoModalProps) {
-  const { t } = useTranslation(['exerciseDataInfo', 'common'])
   // Modal Display Context
-  const [lazyOpen, setLazyOpen] = useState(false)
-  const [fitnessId, setFitnessId] = useState<number | undefined>()
+  const [fitnessId, setFitnessId] = useState<number>(0)
   const contextValue = {
     showModal(fitnessId) {
       setFitnessId(fitnessId)
     }
   } as ModalContextType
 
-  // History Datas
-  const [loadHistory, { data: historyData }] = useLazyGetExerciseFinishHistory()
-  const historyList = useMemo(() => {
-    return (historyData?.getExerciseFinishHistory || []).map((h) => <DisplayFitnessFinishHistory history={h} key={h.id} />)
-  }, [historyData])
-
-  // Fitness Datas
-  const fitnessData = useFitnessFragment(fitnessId ?? 0)
-
-  const fitnessInstructions = fitnessData.instructions
-  const fitnessName = fitnessData.name
-
-  // Fitness Youtube Video
-  const [fitnessVideoId, setFitnessVideoId] = useState<string | undefined>()
-  useEffect(() => {
-    if (fitnessName)
-      FitnessPreviewVideo(fitnessName).then(result => setFitnessVideoId(result))
-    else setFitnessVideoId(undefined)
-  }, [fitnessName])
-
-  useEffect(() => {
-    if (fitnessId) {
-      loadHistory({
-        variables: {
-          exerciseId: fitnessId
-        }
-      })
-    }
-    setLazyOpen(Boolean(fitnessId))
-  }, [fitnessId, loadHistory])
-
   return <ModalContext.Provider value={contextValue}>
     {children}
     <Modal
-      isOpen={lazyOpen}
-      onOpenChange={(v) => {
-        setFitnessId(v ? fitnessId : undefined)
-      }}
+      isOpen={Boolean(fitnessId)}
+      onClose={() => setFitnessId(0)}
       placement='center'
       scrollBehavior='inside'
     >
       <ModalContent>
-        {(onCloseAction) => <>
-          <ModalHeader>
-            {fitnessName || ''}
-          </ModalHeader>
-          <ModalBody>
-            {/* Preview Video */}
-            <StateRender.Boolean
-              state={Boolean(fitnessVideoId)}
-              render={{
-                true: <div >
-                  <iframe
-                    width="100%"
-                    height="315"
-                    sandbox="allow-scripts allow-same-origin allow-presentation"
-                    src={`https://www.youtube.com/embed/${fitnessVideoId}`}
-                    title="YouTube video player"
-                    frameBorder="0"
-                    allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share"
-                    referrerPolicy="strict-origin-when-cross-origin"
-                    allowFullScreen
-                    credentialless="true"
-                  ></iframe>
-                </div>
-              }}
-            />
-            {/* Training History - optional */}
-            <StateRender.Boolean
-              state={Boolean(historyList.length)}
-              render={{
-                true: <div className="max-w-full">
-                  <p className="font-bold text-lg">
-                    {t('history')}
-                  </p>
-                  <ScrollShadow orientation='horizontal'>
-                    {historyList}
-                  </ScrollShadow>
-                </div>
-              }}
-            />
-            {/* Instructions */}
-            <div>
-              <p className="font-bold text-lg">{t('instructions')}</p>
-              <ol className="max-w-full list-decimal list-outside">
-                {fitnessInstructions.map((line, idx) => {
-                  return <li className="ml-6" key={`${fitnessId}-${idx}`}>{line}</li>
-                })}
-              </ol>
-            </div>
-          </ModalBody>
-          <ModalFooter>
-            {/* CloseBtn */}
-            <Button onPress={() => {
-              onCloseAction()
-            }}>
-              {t('common:close')}
-            </Button>
-          </ModalFooter>
-        </>}
+        {(onCloseAction) =>
+          <Suspense>
+            <FitnessDataModalContent fitnessId={fitnessId} onCloseAction={onCloseAction} />
+          </Suspense>
+        }
       </ModalContent>
     </Modal>
   </ModalContext.Provider>
